+++ database/zookeeper/client.go
@@ -13,7 +13,6 @@ import (
 import (
 	jerrors "github.com/juju/errors"
 	"github.com/samuel/go-zookeeper/zk"
-	"github.com/AlexStocks/goext/log"
 )
 
 type Client struct {
@@ -109,7 +108,6 @@ func (c *Client) RegisterTemp(path string, data []byte) (string, error) {
 	}
 
 	tmpPath, err = c.conn.Create(path, data, zk.FlagEphemeral, zk.WorldACL(zk.PermAll))
-	gxlog.CInfo("fuck$$$$$$$$$$$$$$, path:%s, data:%s, err:%s", path, string(path), err)
 	if err != nil {
 		return "", jerrors.Annotatef(err, "zk.Create(%s, ephemeral)", path)
 	}